# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Agentic System API",
    "description": "Sistema Agentic Generalista com Meta-Agent dinamico",
    "version": "v1"
  },
  "paths": {
...
    "/api/admin/settings/gateway": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewaySettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewaySettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GatewaySettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/settings/memory": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/settings/reranking": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReRankingSettingsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReRankingSettingsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReRankingSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GatewaySettings": {
        "type": "object",
        "properties": {
          "defaultDailyBudget": {
            "type": "number",
            "format": "double"
          },
          "defaultFailureThreshold": {
            "type": "integer",
            "format": "int32"
          },
          "defaultBreakDurationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "defaultRequestsPerMinute": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "MemorySettings": {
        "type": "object",
        "properties": {
          "obsidianVaultPath": {
            "type": "string",
            "nullable": true
          },
          "vectorStoreType": {
            "type": "string",
            "nullable": true
          },
          "connectionString": {
            "type": "string",
            "nullable": true
          },
          "qdrant": {
            "$ref": "#/components/schemas/QdrantSettings"
          },
          "pinecone": {
            "$ref": "#/components/schemas/PineconeSettings"
          }
        },
        "additionalProperties": false
      },
...
      "PineconeSettings": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string",
            "nullable": true
          },
          "environment": {
            "type": "string",
            "nullable": true
          },
          "host": {
            "type": "string",
            "nullable": true
          },
          "indexName": {
            "type": "string",
            "nullable": true
          },
          "namespace": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QdrantSettings": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "apiKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateReRankingSettingsRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "useDedicatedProvider": {
            "type": "boolean"
          },
          "dedicatedProvider": {
            "type": "string",
            "nullable": true
          },
          "dedicatedProviderBaseUrl": {
            "type": "string",
            "nullable": true
          },
          "dedicatedProviderModel": {
            "type": "string",
            "nullable": true
          },
          "hasDedicatedProviderApiKey": {
            "type": "boolean"
          },
          "dedicatedProviderTimeoutSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "localOnnxModelPath": {
            "type": "string",
            "nullable": true
          },
          "localOnnxVocabularyPath": {
            "type": "string",
            "nullable": true
          },
          "localOnnxMaxSequenceLength": {
            "type": "integer",
            "format": "int32"
          },
          "localOnnxMaxQueryTokens": {
            "type": "integer",
            "format": "int32"
          },
          "localOnnxLowerCase": {
            "type": "boolean"
          },
          "localOnnxInputIdsName": {
            "type": "string",
            "nullable": true
          },
          "localOnnxAttentionMaskName": {
            "type": "string",
            "nullable": true
          },
          "localOnnxTokenTypeIdsName": {
            "type": "string",
            "nullable": true
          },
          "localOnnxOutputName": {
            "type": "string",
            "nullable": true
          },
          "localOnnxPositiveLabelIndex": {
            "type": "integer",
            "format": "int32"
          },
          "useEmbeddingReRanking": {
            "type": "boolean"
          },
          "useLlmReRanking": {
            "type": "boolean"
          },
          "candidatePoolSize": {
            "type": "integer",
            "format": "int32"
          },
          "minCandidateCountForLlm": {
            "type": "integer",
            "format": "int32"
          },
          "maxSnippetCharacters": {
            "type": "integer",
            "format": "int32"
          },
          "maxOutputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "temperature": {
            "type": "number",
            "format": "float"
          },
          "heuristicConfidenceThreshold": {
            "type": "number",
            "format": "double"
          },
          "heuristicConfidenceGap": {
            "type": "number",
            "format": "double"
          },
          "neuralScoreWeight": {
            "type": "number",
            "format": "double"
          },
          "llmScoreWeight": {
            "type": "number",
            "format": "double"
          },
          "hasUploadedLocalOnnxModel": {
            "type": "boolean"
          },
          "uploadedLocalOnnxModelFileName": {
            "type": "string",
            "nullable": true
          },
          "hasUploadedLocalOnnxVocabulary": {
            "type": "boolean"
          },
          "uploadedLocalOnnxVocabularyFileName": {
            "type": "string",
            "nullable": true
          },
          "dedicatedProviderApiKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Admin API Key",
        "name": "X-Api-Key",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT token com claim tenant_id",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}